requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.4.4
aiohttp==3.10.11
aiosignal==1.3.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.5.2
asgiref==3.8.1
async-timeout==4.0.3
attrs==25.3.0
backoff==2.2.1
backports.zoneinfo==0.2.1
bcrypt==5.0.0
beautifulsoup4==4.14.3
build==1.2.2.post1
certifi==2026.2.25
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.6
chroma-hnswlib==0.7.3
chromadb==0.4.14
click==8.1.8
colorama==0.4.6
coloredlogs==15.0.1
cryptography==46.0.6
dataclasses-json==0.6.7
Deprecated==1.3.1
distro==1.9.0
Django==4.2.29
djangorestframework==3.15.2
durationpy==0.10
emoji==2.15.0
et_xmlfile==2.0.0
eval_type_backport==0.2.2
exceptiongroup==1.3.1
fastapi==0.124.4
filelock==3.16.1
filetype==1.2.0
flatbuffers==25.12.19
frozenlist==1.5.0
fsspec==2025.3.0
googleapis-common-protos==1.73.0
graphlib_backport==1.1.0
greenlet==3.1.1
grpcio==1.70.0
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
huggingface_hub==0.36.2
humanfriendly==10.0
idna==3.11
importlib_metadata==8.5.0
importlib_resources==6.4.5
joblib==1.4.2
jsonpatch==1.33
jsonpath-python==1.1.5
jsonpointer==3.0.0
kubernetes==35.0.0
langchain==0.1.16
langchain-community==0.0.32
langchain-core==0.1.53
langchain-text-splitters==0.0.2
langdetect==1.0.9
langsmith==0.1.147
lxml==6.0.2
markdown-it-py==3.0.0
marshmallow==3.22.0
mdurl==0.1.2
mmh3==5.0.1
monotonic==1.6
mpmath==1.3.0
multidict==6.1.0
mypy_extensions==1.1.0
nest-asyncio==1.6.0
nltk==3.9.1
numpy==1.24.4
oauthlib==3.3.1
ollama==0.6.1
onnxruntime==1.19.2
openpyxl==3.1.5
opentelemetry-api==1.33.1
opentelemetry-exporter-otlp-proto-common==1.33.1
opentelemetry-exporter-otlp-proto-grpc==1.33.1
opentelemetry-instrumentation==0.54b1
opentelemetry-instrumentation-asgi==0.54b1
opentelemetry-instrumentation-fastapi==0.54b1
opentelemetry-proto==1.33.1
opentelemetry-sdk==1.33.1
opentelemetry-semantic-conventions==0.54b1
opentelemetry-util-http==0.54b1
orjson==3.10.15
overrides==7.7.0
packaging==23.2
posthog==2.4.0
propcache==0.2.0
protobuf==5.29.6
pulsar-client==3.5.0
pycparser==2.23
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.19.2
pypdf==5.9.0
PyPika==0.51.1
pyproject_hooks==1.2.0
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-iso639==2024.10.22
python-magic==0.4.27
PyYAML==6.0.3
rapidfuzz==3.9.7
regex==2024.11.6
requests==2.32.4
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.3.3
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.48
sqlparse==0.5.5
starlette==0.44.0
sympy==1.13.3
tabulate==0.9.0
tenacity==8.5.0
tiktoken==0.7.0
tokenizers==0.20.3
tomli==2.4.1
tqdm==4.67.3
typer==0.20.1
typing-inspect==0.9.0
typing_extensions==4.13.2
tzdata==2025.3
unstructured==0.11.8
unstructured-client==0.28.1
urllib3==2.2.3
uvicorn==0.33.0
watchfiles==0.24.0
websocket-client==1.8.0
websockets==13.1
wrapt==1.17.3
yarl==1.15.2
zipp==3.20.2

